+++ bin/src/main.rs
@@ -16,6 +16,7 @@ use ferron_core::registry::{Registry, RegistryBuilder};
 use ferron_core::runtime::Runtime;
 use ferron_core::shutdown::{RELOAD_TOKEN, SHUTDOWN_TOKEN};
 use ferron_core::{log_debug, log_info, log_warn};
+use ferron_observability_prometheus::PrometheusObservabilityModuleLoader;
 use malloc_best_effort::BEMalloc;
 
 use ferron_admin_api::AdminApiModuleLoader;
@@ -283,6 +284,7 @@ fn get_loaders() -> Vec<Box<dyn ModuleLoader>> {
         Box::new(ConsoleObservabilityModuleLoader::default()),
         Box::new(LogFileObservabilityModuleLoader::default()),
         Box::new(OtlpObservabilityModuleLoader::default()),
+        Box::new(PrometheusObservabilityModuleLoader::default()),
         Box::new(JsonFormatObservabilityModuleLoader),
         Box::new(TextFormatObservabilityModuleLoader),
         Box::new(ProcessMetricsModuleLoader::default()),